src/requirements.txt
psutil~=6.1.1
PyQt5~=5.15.11
pyobjc-framework-Quartz~=11.0
pytesseract~=0.3.13
pillow~=11.1.0
numpy~=2.2.2